# perf: reduce one copy behavior during transmission