@doitian 2020-06-04
Can we include the benchmark in CI? Say if there's a 20% performance degradation, the CI fails.